Remove deprecated loadbalancer.
